add open functionality to action menu

+++ src/main/java/dev/candar/swing/JFActionMenu.java
@@ -15,23 +15,54 @@ import javax.swing.event.PopupMenuListener;
 class JFActionMenu  extends JPopupMenu {
     boolean visible = false;
 
+    static enum ActionType {
+        FOLDER,
+        SINGLE_FILE,
+        MULTI_FILES
+    }
 
-    public JFActionMenu() {
-        JMenuItem openItem = new JMenuItemFixed("Open");
-        JMenuItem renameItem = new JMenuItemFixed("Rename");
-        JMenuItem cutItem = new JMenuItemFixed("Cut");
-        JMenuItem copyItem = new JMenuItemFixed("Copy");
-        JMenuItem deleteItem = new JMenuItemFixed("Delete");
-        add(openItem);
-        addSeparator();
-        add(cutItem);
-        add(copyItem);
-        addSeparator();
-        add(renameItem);
-        addSeparator();
-        add(deleteItem);
+    JFFolderViewPanel folderViewPanel;
 
-        
+    JMenuItem newFolderItem = new JMenuItemFixed("New folder");
+    JMenuItem newFileItem = new JMenuItemFixed("New file");
+    JMenuItem openInTerminalItem = new JMenuItemFixed("Open in terminal");
+
+    JMenuItem openItem = new JMenuItemFixed("Open");
+    JMenuItem renameItem = new JMenuItemFixed("Rename");
+    JMenuItem cutItem = new JMenuItemFixed("Cut");
+    JMenuItem copyItem = new JMenuItemFixed("Copy");
+    JMenuItem deleteItem = new JMenuItemFixed("Delete");
+
+    public JFActionMenu(ActionType type) {
+
+        if (type == ActionType.FOLDER) {
+            add(newFolderItem);
+            add(newFileItem);
+            addSeparator();
+            add(openInTerminalItem);
+        }
+
+        if (type == ActionType.SINGLE_FILE) {
+            add(openItem);
+            addSeparator();
+            add(cutItem);
+            add(copyItem);
+            addSeparator();
+            add(renameItem);
+            addSeparator();
+            add(deleteItem);
+        }
+
+        if (type == ActionType.MULTI_FILES) {
+            add(cutItem);
+            add(copyItem);
+            addSeparator();
+            add(deleteItem);
+        }
+
+        openItem.addActionListener((action) -> {
+            folderViewPanel.openSelectedFile();
+        });
         
     }
 }
